feat: subcomponent linking
Introduce a `SubComponent` attribute on links that will replace the
use of component type specific link attributes. For now, either will
work.

Also expand the checking of component types and the use of the `Type`
attribute, particularly as a check on the type of the expected link
target and to differentiate between subcomponents with the same id
but different types.