fix: qualification of layer data
Permit the use of qualification within layer data.

Rework the handling of layers to permit their data to provide the
filter inputs (Id, Name) to their own qualification.

Layer data is processed once to work out the values to use when
filtering it, then a second time to work the effects of qualification
on the remaining configuration data.

As part of the refactoring, the active state of layers in now part of
the input state and calls to `getActiveLayers` now returns the data in
the input state.