refactor: clo/layer access during input processing
Refactor input processing to permit access to command line options and
layers during input processing. For some stages of input stages, the
state to that point is saved into the global state, and input state
accessor routines have been made aware of whether input processing is
in progress in order to use this staged state.

Access to layer information is commonly used in extensions and modules.