fix: permit the mocking of empty values
Permit the use of a "marker" value of `hamlet:empty` to signify that the
value of a stack output should be an empty string. This is used during
testing to ensure that legacy stack outputs can be simulated. The empty
value is treated as a missing output rather than always being present.

The marker value is converted to lower case so any preference for casing
can be used - all lower case recommended.

Correct several bugs related to the handling of fixture and mock stack
outputs, mainly related to appending rather than merging stack outputs.

Note that when providing mock stack outputs, Account, Region, and
DeploymentUnit must also be provided e.g.

```
stackOutputs=[
    {
        "Account" : "0123456789",
        "Region" : "mock-region-1",
        "DeploymentUnit" : "aws-cdn-spa-base",
        "cfXwebXcdnspabase": "hamlet:empty"
    }
]
```