refactor: wildcard behaviour of getCompositeObject (#1786)
At present, the name wildcard processing of `getCompositeObject()` uses
freemarker object addition. This means that only one object can
contribute the content for any top level attribute.

This change replaces this with the use of `mergeObjects()`. The change
will only affect top level attributes with an object value. The effect
is that multiple input objects can contribute to the overall value.

The driver for the change is the use of modules to provide default
settings content which can then be supplemented with explicit settings,
an example being the provision of default api gateway integrations in a
module.

A small documentation update for openapi processing is also included.