refactor: remove last of component qualifiers
Remove the specific `Qualifiers` attributes that were still in place
for certificate related activities.

When global qualification was added to the input processor, these
attributes were left in place as a precaution. However global
qualification means they will be removed before seen as part as of the
component configuration.

The use of the `Qualifiers` attribute has also been deprecated in favour
of the namespaced qualifier attributes `qualifier:Defualt` and
`qualifier:Rules` so seems like the right time to clean up these
attributes.

The change also includes removal of some experimental code that has been
superceded by the implementation of the cmdb provider.