feat: subscription, hostingplatform, dnszone components
Add support for new `subscription`, `hostingplatform` and `dnszone` components.

Part of [ADR0007](https://github.com/hamlet-io/architectural-decision-log/blob/main/adr/0007-placement-support.md) implementation

- Add shared definition of these new components
- Add preliminary changes to support `Locations`
  - accessor methods for location information on components
and occurrences
  - ability to capture the errors generated by `getCompositeObject()`
so alternate actions to generating fatal errors can be supported,
mainly so the context of the errors can be better explained
  - More extensive commenting of the process for attribute extension
and some refactoring of the implementation
- new component attributes support import of externally
defined resources in readiness for the introduction of `Locations`.
- Support for an explicit Type attribute on components rather than
relying on the name of a child attribute
  - makes config more explicit
  - permits a flatter JSON file
  - in line with similar changes to links with the removal of
subcomponent specific attributes

Subsequent PRs will add explicit support for Locations.