feat(locations): location support
Add support for Locations on ResourceGroups as a means to determine
the placement of resources.

When determining placements, the resource group id is used
as the Location to find the corresponding placement information,
so this means the ids of all resource groups need to be known
BEFORE location information is checked.

As a result, when adding a component, any possible
resource groups that an implementation might want to use must be
defined. This means that a provider implementer cannot add
additional resource groups not already defined when the component
was defined.

The existing placement profile approach has been left in place
but if present, locations take precedence.

As locations are link based, their introduction increases the
chances of creating loops. Code to detect loops has thus
been added when getting Occurrences.