refactor: enforce resource group registration
All resource groups for a component must now be registered at the time
the component is registered. An attempt to add information for an
unregistered resource group will result in an error.

This change is needed to ensure it is possible to determine resource
group placement BEFORE knowing the placement provider.