refactor: occurrence logging
Add a helper function to decouple what information in an occurrence
gets logged in log messages from where the information is needed.

The size of the occurrence structure means it does clutter logs if
provided in full, and the need for the full set of data is rare.

By decoupling the two, additional flexibility around what gets logged
can be added without having to revisit all the places where occurrence
information is included in logs.