refactor: permit transition for fragments
Fragments and extensions have been written on the assumption of the
availability of the global variables defined by `setContext`. To
permit a transition away from this, move the definition of the
global variables to a single macro and for each variable, add an
accessor function starting with `get...`.

All internal uses of the global variables are to be converted to uses
of the accessors, with the naming making the process easy to review from
a change perspective.

The macro to set the variables is only invoked in `invokeExtensions`
and have an argument to permit selective disabling of the variables in
a subsequent phase where the default will switch to the globals NOT
being defined by able to be selectively enabled to permit
flexibility on the upgrade of existing fragments/extensions.