refactor: pseudo stack outputs (#1869)
Permit the override of the base output points.

Treat outputs provided by the caller as of higher priority than the
base set of points added by default (account, region, unit, mode).

This permits the (admittedly rare) situation when the default values
are not appropriate. In most cases, the caller assumes the function
will add these values.