feat: LinkRef support for container LB references
Add support for the use of LinkRefs to connect containers to
the load balancers against which they need to register.

The support for using link lookup to perform this connection was
already in place. The changes in this commit simply add LinkRef
processing as part of determining the link definition to use
in the lookup.