feat(backup): Backup support
Add a component representing a store for backups sourced from
one or more components.

Regimes are used to determine combinations of schedule and component
for which snapshots should be generated and saved in the store.

Components to be backed up can be selected via
- carrying tags that match those of the backup store
- carrying a tag specific to a regime on a backup store
- direct links from the backup store.

Copies of snapshots can be saved in additional backup stores if
required.