refactor(apigateway): authorization models (#1995)
* refactor(apigateway): authorization models

Support additional values for the case of IP filtering in combination
with a lambda authorizer. Also rename the config attribute to
more correctly reflect its purpose in controlling authorization
rather than authentication.

When used with a lambda authorizer, the default value of "IP"
incorrectly provides an explict ALLOW  rather than relying on it
to come from the policy provided by the authorizer. By providing
explicit values to be used with the authorizer, the configuration
can be validated as appropriate.

* chore(apigateway): add description

Add description of the AuthorisationModel attribute.

* fix(apigateway): remove SIG4 models

As SIG4 is very AWS specific, move all SIG4 related values to the
AWS provider. The non-prefixed variant will still be included
for backwards compatability.

With the new prefixed AWS values in place, a future breaking change can
remove the non-prefixed SIG4 specific values from the AWS provider.