feat(apigateway): private APIs (#2115)
Add the option of private APIs. For these APIS, traffic filtering is
done on the basis of VPC endpoints rather than IP addresses. The VPC
endpoints are assumed to be configured via inbound links to external
services, which should provide a VPC_ENDPOINTS attribute containing a
command separated list of vpc endpoint names.

Adjust the available values for the authorisation model to support
either IPs or VPC Endpoints.